Feat: evolving frontend appearance
